requirements.txt
# Основные зависимости
pytest>=7.0.0
selenium>=4.0.0
requests>=2.28.0
allure-pytest>=2.12.0
python-dotenv>=1.0.0   
webdriver-manager>=4.0.0